Working from dejuknow's mavericks-list-view branch, the TDimmableIconImageView draRect will now extract the file path from the FINode living as an instance variable on the super super view, which is a TListRowView, and use it to select the correct image from the IconCache.
The only obvious remaining todo is to figure out how to get the list view to repaint. I've tried invalidating the entire tree of subviews, which does not seem to work. Inspection via GDB suggests that the list view is backed by CALayer, which probably needs to be invalidated in some different manner.